Moved BaseCommand class to abstracts dir
